AIservices/requirements.txt
fastapi
uvicorn
langchain
langchain-core
langchain-community
langchain-google-genai
langchain-groq
huggingface_hub
python-dotenv
pydantic
requests
chromadb
sentence-transformers
